+++ packages/core-js-compat/src/external.js
@@ -1,0 +1,9 @@
+'use strict';
+module.exports = {
+  modules: {
+    edge: '16',
+    chrome: '61',
+    firefox: '60',
+    safari: '10.1',
+  },
+};