+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 ##### Unreleased
-- Allowed object of minimum environment versions as `core-js-compat/compat` argument
+- Allowed object of minimum environment versions as `core-js-compat` and `core-js-builder` `targets` argument
+- Allowed corresponding to Babel `targets.esmodules`, `targets.browsers`, `targets.node` options in `core-js-compat` and `core-js-builder`
 - Added Samsung Internet 11 compat data mapping
 
 ##### 3.5.0 - 2019.12.12